local_requirements.txt
beam-client
gunicorn
flask